requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
aiohttp
pdfplumber
sentence-transformers
faiss-cpu
numpy
pydantic
requests
